fix(auth): ensure admin session cookie is correctly scoped and recognized
- Updated set_cookie and delete_cookie to explicitly set domain to 10.1.2.97
- Confirmed session cookie is now shared between frontend and backend
- Added logging in check-session to debug cookie alias mapping
- Ensured frontend fetch uses credentials: 'include' for session persistence
- Aligned cookie name alias with actual browser cookie key
