+++ App.py
@@ -15,7 +15,10 @@ from utils.logger import setup_logging
 from utils.file_ops import create_upload_dir
 from utils.cleanup import cleanup_old_batches, cleanup_temp_uploads, log_cleanup_stats
 from utils.ws_manager import connection_manager
-from routers import single, batch, export, admin_auth, batch_history
+from utils.security import csrf_protection
+# Import admin_auth first to avoid circular imports
+from routers import admin_auth
+from routers import single, batch, export, batch_history
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 import logging
 import asyncio
@@ -41,20 +44,28 @@ app.add_exception_handler(RateLimitExceeded, _rate_limit_exceeded_handler)
 async def log_requests(request: Request, call_next):
     """Log all incoming HTTP requests and their responses"""
     logger.info(f"Request: {request.method} {request.url}")
+    
+    # Debug cookies for admin endpoints
+    if "/api/admin/" in str(request.url):
+        logger.info(f"Request cookies: {request.cookies}")
+    
     response = await call_next(request)
+    
     logger.info(
         f"Response: {request.method} {request.url} - Status: {response.status_code}"
     )
+    
     return response
 
 
 # Configure CORS for cross-origin requests
 app.add_middleware(
     CORSMiddleware,
-    allow_origins=["http://localhost:3000"],
+    allow_origins=["http://localhost:3000", "http://10.1.2.97:3000"],
     allow_credentials=True,
     allow_methods=["*"],
     allow_headers=["*"],
+    expose_headers=["X-CSRF-Token", "Content-Type", "Content-Disposition", "X-Total-Count"],
 )
 
 
@@ -89,6 +100,10 @@ async def startup_event():
         app.state.scheduler.add_job(cleanup_old_batches, "interval", hours=1, args=[24])
         app.state.scheduler.add_job(
             cleanup_temp_uploads, "interval", minutes=30, args=[30]
+        )
+        # Add CSRF token cleanup job
+        app.state.scheduler.add_job(
+            csrf_protection.clean_expired_tokens, "interval", minutes=15
         )
 
         # Start the scheduler
