+++ routers/export.py
@@ -1,10 +1,14 @@
-from fastapi import APIRouter, HTTPException
+from fastapi import APIRouter, HTTPException, Header, Query
 from starlette.requests import Request
 from fastapi.responses import FileResponse
 from slowapi import Limiter
 from slowapi.util import get_remote_address
+from routers.batch import check_token_valid
 from utils.cleanup import cleanup_old_batches, cleanup_temp_uploads, log_cleanup_stats
+from typing import Optional
+import sys
 import logging
+
 import os
 import json
 
@@ -16,7 +20,7 @@ router = APIRouter(tags=["Export"])
 
 @router.get("/check-logo/batch/getCount")
 async def get_last_batch_count(request: Request, batch_id: str):
-    batch_dir = os.path.join("data", batch_id)
+    batch_dir = os.path.join("exports", batch_id)
     metadata_path = os.path.join(batch_dir, "metadata.json")
     if not os.path.exists(metadata_path):
         raise HTTPException(status_code=400, detail=f"Invalid batch_id: {batch_id}")
@@ -35,7 +39,7 @@ async def export_batch_results_csv(request: Request, batch_id: str):
     Returns a downloadable CSV file with the results.
     """
     try:
-        batch_dir = os.path.join("data", batch_id)
+        batch_dir = os.path.join("exports", batch_id)
         metadata_path = os.path.join(batch_dir, "metadata.json")
         if not os.path.exists(metadata_path):
             raise HTTPException(status_code=400, detail=f"Invalid batch_id: {batch_id}")
@@ -45,10 +49,13 @@ async def export_batch_results_csv(request: Request, batch_id: str):
 
         csv_path = metadata["csv_path"]
 
+        filename = f"logo_detection_results_{batch_id}.csv"
+        # Ensure filename has no extra spaces
+        filename = filename.strip()
         return FileResponse(
             csv_path,
             media_type="text/csv",
-            filename=f"logo_detection_results_{batch_id}.csv",
+            filename=filename,
         )
 
     except Exception as e:
@@ -56,15 +63,52 @@ async def export_batch_results_csv(request: Request, batch_id: str):
         raise HTTPException(status_code=500, detail=str(e))
 
 
+@router.get("/api/exports/{filename}")
+async def get_export_file(
+    filename: str, 
+    token: Optional[str] = Query(None)
+):
+    """
+    Get an exported file by filename.
+    Requires authentication token as a query parameter.
+    """
+    try:
+        # Validate token
+        if not token or not check_token_valid(token):
+            raise HTTPException(status_code=401, detail="Authentication required")
+        
+        file_path = os.path.join("exports", filename)
+        if not os.path.exists(file_path):
+            raise HTTPException(status_code=404, detail=f"File not found: {filename}")
+        
+        return FileResponse(
+            file_path,
+            media_type="text/csv",
+            filename=filename
+        )
+    except HTTPException:
+        raise
+    except Exception as e:
+        logger.error(f"Error serving export file: {str(e)}")
+        raise HTTPException(status_code=500, detail=str(e))
+
+
 @router.post("/maintenance/cleanup")
 @limiter.limit("2/minute")
 async def trigger_cleanup(
-    request: Request, batch_age_hours: int = 24, temp_age_minutes: int = 30
+    request: Request, 
+    batch_age_hours: int = 24, 
+    temp_age_minutes: int = 30,
+    token: Optional[str] = Header(None, alias="X-Auth-Token")
 ):
     """
     Manually trigger cleanup of old batch results and temporary files.
     Returns the number of items cleaned up.
     """
+    # Validate token for admin operations
+    if not token or not check_token_valid(token):
+        raise HTTPException(status_code=401, detail="Authentication required")
+    
     try:
         batch_cleaned = cleanup_old_batches(max_age_hours=batch_age_hours)
         temp_cleaned = cleanup_temp_uploads(max_age_minutes=temp_age_minutes)
